requirements.txt
pillow
piexif
tqdm